Move Application class to its own module (#22)
* Move Application class to its own module

- separate the Application class from the analyze
- makes easier the review on next patch containing changes on this
  class because github doesn't recognize when a file is moved.
- changed location to cou/apps
